Refactor AI integration: replace GeminiProvider with AIProvider and implement new AI features
