+++ src/main/kotlin/me/voxelsquid/quill/event/UniqueItemGenerateEvent.kt
@@ -1,12 +1,12 @@
 package me.voxelsquid.quill.event
 
-import me.voxelsquid.quill.ai.GeminiProvider
+import me.voxelsquid.quill.ai.AIProvider
 import org.bukkit.entity.Villager
 import org.bukkit.event.Event
 import org.bukkit.event.HandlerList
 import org.bukkit.inventory.ItemStack
 
-class UniqueItemGenerateEvent(val villager: Villager, val item: ItemStack, val data: GeminiProvider.UniqueItemDescription) :
+class UniqueItemGenerateEvent(val villager: Villager, val item: ItemStack, val data: AIProvider.UniqueItemDescription) :
     Event() {
 
     override fun getHandlers(): HandlerList {
